agent/requirements.txt
websockets>=12.0
psutil>=5.9.8
pyyaml>=6.0.1
requests>=2.31.0
pydantic>=2.6.4
asyncio>=3.4.3
